+++ api/models/User.js
@@ -5,7 +5,7 @@
 * @docs        :: http://sailsjs.org/#!documentation/models
 */
 
-var bcrypt = require( 'bcrypt' );
+var bcrypt = require( 'bcrypt-nodejs' );
 var uniqueNickname = false;
 
 module.exports = {
@@ -79,6 +79,7 @@ module.exports = {
 		bcrypt.hash(
 			value.password,
 			10,
+			function() { /*progress*/ },
 			function passwordEncrypted( err, encryptedPassword ) {
 				if( err ) return next( err );
 				value.password = encryptedPassword;