+++ config/env/lite.js
@@ -1,0 +1,12 @@
+/**
+ * Quirkbot Lite environment settings
+ *
+ */
+
+module.exports = {
+
+  models: {
+    connection: 'lite'
+  }
+
+};